+++ src/main/java/com/oti/srm/controller/notice/NoticeController.java
@@ -1,4 +1,4 @@
-package com.oti.srm.controller;
+package com.oti.srm.controller.notice;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
