+++ src/main/java/com/oti/srm/controller/srm/RequestController.java
@@ -1,4 +1,4 @@
-package com.oti.srm.controller;
+package com.oti.srm.controller.srm;
 
 import java.util.List;
 
@@ -6,17 +6,15 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.oti.srm.dto.Member;
 import com.oti.srm.dto.Pager;
 import com.oti.srm.dto.Request;
-import com.oti.srm.service.IRequestRegisterService;
-import com.oti.srm.service.IUserRegisterService;
+import com.oti.srm.service.member.IUserRegisterService;
+import com.oti.srm.service.srm.IRequestRegisterService;
 
 import lombok.extern.log4j.Log4j2;
