+++ src/main/java/com/oti/srm/dao/srm/IRequestDao.java
@@ -1,4 +1,4 @@
-package com.oti.srm.dao;
+package com.oti.srm.dao.srm;
 
 import java.util.List;
