+++ src/main/java/com/oti/srm/service/member/IUserRegisterService.java
@@ -1,4 +1,4 @@
-package com.oti.srm.service;
+package com.oti.srm.service.member;
 
 import com.oti.srm.dto.Member;
 
@@ -6,9 +6,10 @@ public interface IUserRegisterService {
 	public enum RegisterResult {
 		SUCCESS, FAIL;
 	}
+
 	public static final int REGISTER_SUCCESS = 0;
 	public static final int REGISTER_FAIL = 1;
-	
+
 	public int register(Member member);
 
 }
