+++ src/main/java/com/oti/srm/service/srm/UserTestService.java
@@ -1,19 +1,20 @@
-package com.oti.srm.service;
+package com.oti.srm.service.srm;
 
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.oti.srm.dao.ICommonDao;
+import com.oti.srm.dao.srm.ICommonDao;
 import com.oti.srm.dto.Request;
 import com.oti.srm.dto.RequestProcess;
 import com.oti.srm.dto.StatusHistory;
 
 @Service
-public class DistributeService implements IDistributeService {
+public class UserTestService implements IUserTestService {
 	@Autowired
 	ICommonDao commonDao;
 
@@ -48,6 +49,26 @@ public class DistributeService implements IDistributeService {
 	public RequestProcess getRequestProcess(int rno) {
 		RequestProcess requestProcess = commonDao.getRequestProcess(rno);
 		return requestProcess;
+	}
+
+	// 작업 시작(고객테스터 / 배포자 공용)
+	// => requests테이블(현재단계 최신화 + 완료예정일 기입) + status_histories테이블(단계 변경 이력 추가)
+	@Override
+	@Transactional
+	public void startWork(StatusHistory statusHistory, Date expectDate, String mtype) {
+		commonDao.updateExpectDate(statusHistory.getRno(), expectDate, mtype);
+		commonDao.updateRequestStatus(statusHistory.getRno(), statusHistory.getNextStatus());
+		commonDao.insertStatusHistory(statusHistory);
+	}
+
+	// 작업 완료(고객테스터 / 배포자 공용)
+	// => requests테이블(현재단계 최신화) + status_histories테이블(단계 변경 이력 추가) 
+	// * 파일이 있다면 status_histories_files테이블(단계 변경 이력에 첨부파일 등록)
+	@Override
+	@Transactional
+	public void endWork(StatusHistory statusHistory) {
+		commonDao.updateRequestStatus(statusHistory.getRno(), statusHistory.getNextStatus());
+		commonDao.insertStatusHistory(statusHistory);
 	}
 
 }
