+++ src/main/resources/mybatis/mapper/common.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.oti.srm.dao.ICommonDao">
+<mapper namespace="com.oti.srm.dao.srm.ICommonDao">
 	<select id="getRequestHistories" parameterType="int" resultType="com.oti.srm.dto.StatusHistory">
 		select
 	    	hno as "hno",
