+++ src/main/resources/mybatis/mapper/request.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.oti.srm.dao.IRequestDao">
+<mapper namespace="com.oti.srm.dao.srm.IRequestDao">
 	<insert id="request" parameterType="com.oti.srm.dto.Request" >
 		insert into requests (RNO, REQ_DATE, REQ_TITLE, REQ_CONTENT, CLIENT, SNO, REQ_EXPECT_DATE, STATUS_NO)
 		values(SEQ_REQUESTS.nextval, sysdate, #{reqTitle}, #{reqContent}, #{client}, #{sno}, #{reqExpectDate}, #{statusNo})
