+++ src/main/resources/mybatis/mapper/userregister.xml
@@ -1,10 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.oti.srm.dao.IUserRegisterDao">
+<mapper namespace="com.oti.srm.dao.member.IUserRegisterDao">
 	<insert id="register" parameterType="com.oti.srm.dto.Member" >
 		insert into members (MID, MTYPE, MNAME, BIRTH, GENDER, ADDRESS, SNO, EMAIL, PASSWORD, ORGAN, PHONE, POSITION)
 		values (#{mid}, '고객', #{mname}, sysdate, #{gender}, #{address}, #{sno}, #{email}, #{password}, #{organ}, #{phone}, #{position})
-		
 	</insert>
